Extract the pattern overrides toolbar indicator from the `block-editor` package (#62514)
* Extract pattern overrides toolbar indicator from the block-editor package

* Fix error
